requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0
pillow>=10.2.0
reportlab>=4.0.0
